+++ src/store/s3/aws_helpers.rs
@@ -1,0 +1,56 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use hmac::{Hmac, Mac};
+use sha2::digest::FixedOutput;
+use sha2::{Digest, Sha256};
+
+/// Generates a signing key for Sigv4
+pub fn generate_signing_key(
+    secret: &str,
+    date: &str,
+    region: &str,
+    service: &str,
+) -> impl AsRef<[u8]> {
+    // kSecret = your secret access key
+    // kDate = HMAC("AWS4" + kSecret, Date)
+    // kRegion = HMAC(kDate, Region)
+    // kService = HMAC(kRegion, Service)
+    // kSigning = HMAC(kService, "aws4_request")
+
+    let secret = format!("AWS4{}", secret);
+    let mut mac =
+        Hmac::<Sha256>::new_from_slice(secret.as_ref()).expect("HMAC can take key of any size");
+    mac.update(date.as_bytes());
+    let tag = mac.finalize_fixed();
+
+    // sign region
+    let mut mac = Hmac::<Sha256>::new_from_slice(&tag).expect("HMAC can take key of any size");
+    mac.update(region.as_bytes());
+    let tag = mac.finalize_fixed();
+
+    // sign service
+    let mut mac = Hmac::<Sha256>::new_from_slice(&tag).expect("HMAC can take key of any size");
+    mac.update(service.as_bytes());
+    let tag = mac.finalize_fixed();
+
+    // sign request
+    let mut mac = Hmac::<Sha256>::new_from_slice(&tag).expect("HMAC can take key of any size");
+    mac.update("aws4_request".as_bytes());
+    mac.finalize_fixed()
+}
+
+pub fn calculate_signature(signing_key: impl AsRef<[u8]>, string_to_sign: &[u8]) -> String {
+    let mut mac = Hmac::<Sha256>::new_from_slice(signing_key.as_ref())
+        .expect("HMAC can take key of any size");
+    mac.update(string_to_sign);
+    hex::encode(mac.finalize_fixed())
+}
+
+pub fn sha256_sum(bytes: impl AsRef<[u8]>) -> String {
+    let mut hasher = Sha256::new();
+    hasher.update(bytes);
+    hex::encode(hasher.finalize_fixed())
+}
